+++ changelog.md
@@ -2,6 +2,12 @@
 ## New features
 * matrix: New protocol support added (https://matrix.org)
 
+## Bugfix
+* mattermost: Add ReadTimeout to close lingering connections (mattermost). See #125
+
+## Changes
+* telegram: Do not use HTML parsemode by default. Set ```MessageFormat="HTML"``` to use it. Closes #126
+
 # v0.9.3
 ## New features
 * API: rest interface to read / post messages (see API section in matterbridge.toml.sample)